Simplified traits.
