+++ src/array/specification.rs
@@ -1,61 +1,4 @@
-use std::convert::TryFrom;
-
-use num_traits::Num;
-
-use crate::types::Index;
-
-mod private {
-    pub trait Sealed {}
-
-    impl Sealed for i32 {}
-    impl Sealed for i64 {}
-}
-
-/// Sealed trait describing types that can be used as offsets in Arrow (`i32` and `i64`).
-pub trait Offset: private::Sealed + Index + Num + Ord + num_traits::CheckedAdd {
-    /// Whether it is `i32` or `i64`
-    fn is_large() -> bool;
-
-    /// converts itself to `isize`
-    fn to_isize(&self) -> isize;
-
-    /// converts from `isize`
-    fn from_isize(value: isize) -> Option<Self>;
-}
-
-impl Offset for i32 {
-    #[inline]
-    fn is_large() -> bool {
-        false
-    }
-
-    #[inline]
-    fn from_isize(value: isize) -> Option<Self> {
-        Self::try_from(value).ok()
-    }
-
-    #[inline]
-    fn to_isize(&self) -> isize {
-        *self as isize
-    }
-}
-
-impl Offset for i64 {
-    #[inline]
-    fn is_large() -> bool {
-        true
-    }
-
-    #[inline]
-    fn from_isize(value: isize) -> Option<Self> {
-        Self::try_from(value).ok()
-    }
-
-    #[inline]
-    fn to_isize(&self) -> isize {
-        *self as isize
-    }
-}
+use crate::types::Offset;
 
 pub fn check_offsets_minimal<O: Offset>(offsets: &[O], values_len: usize) -> usize {
     assert!(
