+++ src/compute/arithmetics/basic/rem.rs
@@ -2,7 +2,7 @@ use std::ops::Rem;
 
 use num_traits::{CheckedRem, NumCast};
 
-use crate::datatypes::DataType;
+use crate::datatypes::PrimitiveType;
 use crate::{
     array::{Array, PrimitiveArray},
     compute::{
@@ -98,8 +98,8 @@ where
 {
     let rhs = *rhs;
 
-    match T::DATA_TYPE {
-        DataType::UInt64 => {
+    match T::PRIMITIVE {
+        PrimitiveType::UInt64 => {
             let lhs = lhs.as_any().downcast_ref::<PrimitiveArray<u64>>().unwrap();
             let rhs = rhs.to_u64().unwrap();
 
@@ -114,7 +114,7 @@ where
                 ))
             }
         }
-        DataType::UInt32 => {
+        PrimitiveType::UInt32 => {
             let lhs = lhs.as_any().downcast_ref::<PrimitiveArray<u32>>().unwrap();
             let rhs = rhs.to_u32().unwrap();
 
@@ -129,7 +129,7 @@ where
                 ))
             }
         }
-        DataType::UInt16 => {
+        PrimitiveType::UInt16 => {
             let lhs = lhs.as_any().downcast_ref::<PrimitiveArray<u16>>().unwrap();
             let rhs = rhs.to_u16().unwrap();
 
@@ -144,7 +144,7 @@ where
                 ))
             }
         }
-        DataType::UInt8 => {
+        PrimitiveType::UInt8 => {
             let lhs = lhs.as_any().downcast_ref::<PrimitiveArray<u8>>().unwrap();
             let rhs = rhs.to_u8().unwrap();
