+++ src/compute/cast/primitive_to.rs
@@ -33,7 +33,7 @@ pub fn primitive_to_binary<T: NativeType + lexical_core::ToLexical, O: Offset>(
             let len = lexical_core::write_unchecked(*x, bytes).len();
 
             offset += len;
-            offsets.push(O::from_isize(offset as isize).unwrap());
+            offsets.push(O::from_usize(offset as usize).unwrap());
         }
         values.set_len(offset);
         values.shrink_to_fit();
@@ -99,7 +99,7 @@ pub fn primitive_to_utf8<T: NativeType + lexical_core::ToLexical, O: Offset>(
             let len = lexical_core::write_unchecked(*x, bytes).len();
 
             offset += len;
-            offsets.push(O::from_isize(offset as isize).unwrap());
+            offsets.push(O::from_usize(offset as usize).unwrap());
         }
         values.set_len(offset);
         values.shrink_to_fit();
