+++ src/compute/sort/lex_sort.rs
@@ -179,8 +179,9 @@ pub fn lexsort_to_indices<I: Index>(
         values.sort_unstable_by(lex_comparator);
     }
 
+    let data_type = I::PRIMITIVE.into();
     Ok(PrimitiveArray::<I>::from_data(
-        I::DATA_TYPE,
+        data_type,
         values.into(),
         None,
     ))
