+++ src/compute/sort/mod.rs
@@ -357,7 +357,8 @@ where
 
     values.truncate(limit.unwrap_or_else(|| values.len()));
 
-    PrimitiveArray::<I>::from_data(I::DATA_TYPE, values.into(), None)
+    let data_type = I::PRIMITIVE.into();
+    PrimitiveArray::<I>::from_data(data_type, values.into(), None)
 }
 
 /// Compare two `Array`s based on the ordering defined in [ord](crate::array::ord).
